requirements.txt
python-gnupg==0.4.6
PyYAML==5.3.1
